Initial commit
Checking in my highlight.vim, basically.

There are still a few todos, but I've cleaned up the latter half of the
file quite a bit (everything after the SYNTAX PALETTE section).

I'm checking it in so I can immediately update my Vim config to use this
as a plugin under ~/.vim/pack.